refactor: 전체적인 레이아웃 구조 변경
