+++ src/main.jsx
@@ -3,22 +3,24 @@ import { BrowserRouter, Routes, Route } from "react-router";
 import "./index.css";
 
 // Layout
-import Header from "./components/Header";
-import Footer from "./components/footer";
+import MainLayout from "./components/layout/MainLayout";
 
 // Pages
 import Home from "./pages/Home";
 import LoginPage from "./pages/Login";
 import SignupPage from "./pages/Signup";
+import ItemsPage from "./pages/ItemsPage";
 
 createRoot(document.getElementById("root")).render(
   <BrowserRouter>
-    <Header />
     <Routes>
-      <Route index element={<Home />} />
-      <Route path="login" element={<LoginPage />} />
-      <Route path="signup" element={<SignupPage />} />
+      <Route element={<MainLayout />}>
+        <Route index element={<Home />} />
+        <Route path="items" element={<ItemsPage />} />
+      </Route>
+
+      <Route path="/login" element={<LoginPage />} />
+      <Route path="/signup" element={<SignupPage />} />
     </Routes>
-    <Footer />
   </BrowserRouter>,
 );
